Adding cardDetails page (done)
incrementing product quantity in cardDetails page (done)

+++ src/main/java/com/jchaaban/cmsshoppingcard/services/CardService.java
@@ -18,7 +18,7 @@ public class CardService {
 
     public void updateCardStatus(HttpSession session, Model model){
 
-        int cardSize = 0, cardTotal = 0;
+        double cardSize = 0, cardTotal = 0;
 
         HashMap<Integer, CardItem> card = (HashMap<Integer, CardItem>) session.getAttribute("card");
